move method from static pages helper to application helper
