+++ spec/helpers/application_helper_spec.rb
@@ -16,4 +16,21 @@ RSpec.describe ApplicationHelper, type: :helper do
       end
     end
   end
+
+  describe '#home_link' do
+    context 'when user is logged in' do
+      it 'returns the user path' do
+        user = create(:user)
+        @current_user = user
+
+        expect(helper.home_link).to eq(user_path(user))
+      end
+    end
+
+    context 'when user is not logged in' do
+      it 'returns the root path' do
+        expect(helper.home_link).to eq(root_path)
+      end
+    end
+  end
 end
